backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "sqlalchemy>=2.0.0",
    "bcrypt>=4.0.0",
    "pydantic>=2.0.0",
    "motor>=3.7.1",
    "pycryptodome>=3.20.0",
    "python-dotenv>=1.0.0",
    "python-jose[cryptography]>=3.3.0",
    "slowapi>=0.1.9",
]
